move BasicAuthenticationProviderConfigurationSchema to security module
